[IMP] spreadsheet: remove unused `PivotUIGlobalFilterPlugin`
Thie plugin `PivotUIGlobalFilterPlugin` was used in the past to set the
global filter value when clicking on a pivot cell, but this is now
done by the `clickableCells`, the code of the plugin is not called
anymore.

Some tests were testing the internal methods of the plugin. We'll
move the plugin's internal code to a test helper, the tests are still
useful as they test the `getFiltersMatchingPivotArgs` getter, that is
used in clicable cells.

Some tests were removed as they either
1) were now just testing the test helper
2) were wrong in the first place, testing that the correct fitlers are
returned, but their actual `expect` expected an empty array

Task: 6105514
