+++ addons/spreadsheet/static/src/pivot/index.js
@@ -5,7 +5,6 @@ import * as spreadsheet from "@odoo/o-spreadsheet";
 import { SEE_RECORDS_PIVOT, SEE_RECORDS_PIVOT_VISIBLE } from "./pivot_actions";
 import { PivotOdooCorePlugin } from "./plugins/pivot_odoo_core_plugin";
 import { PivotCoreViewGlobalFilterPlugin } from "./plugins/pivot_core_view_global_filter_plugin";
-import { PivotUIGlobalFilterPlugin } from "./plugins/pivot_ui_global_filter_plugin";
 
 const { coreTypes, invalidateEvaluationCommands } = spreadsheet;
 
@@ -38,4 +37,4 @@ cellMenuRegistry.add("pivot_see_records", {
 
 inverseCommandRegistry.add("UPDATE_ODOO_PIVOT_DOMAIN", identity);
 
-export { PivotOdooCorePlugin, PivotCoreViewGlobalFilterPlugin, PivotUIGlobalFilterPlugin };
+export { PivotOdooCorePlugin, PivotCoreViewGlobalFilterPlugin };
